+++ AmazonChimeSDKDemo/AmazonChimeSDKDemo/VideoModel.swift
@@ -18,17 +18,44 @@ class VideoModel: NSObject {
     private var remoteVideoTileStates: [(Int, VideoTileState)] = []
     private var userPausedVideoTileIds: Set<Int> = Set()
     public var remoteVideoSourceConfigurations: Dictionary<RemoteVideoSource, VideoSubscriptionConfiguration> = Dictionary()
-    private let audioVideoFacade: AudioVideoFacade
+    let audioVideoFacade: AudioVideoFacade
     let customSource: DefaultCameraCaptureSource
 
     var videoUpdatedHandler: (() -> Void)?
+    var videoSubscriptionUpdatedHandler: (() -> Void)?
     var localVideoUpdatedHandler: (() -> Void)?
     let logger = ConsoleLogger(name: "VideoModel")
+
+    private let backgroundBlurProcessor: BackgroundBlurVideoFrameProcessor
+    private var backgroundReplacementProcessor: BackgroundReplacementVideoFrameProcessor
+    private var backgroundImage: UIImage?
 
     init(audioVideoFacade: AudioVideoFacade, eventAnalyticsController: EventAnalyticsController) {
         self.audioVideoFacade = audioVideoFacade
         self.customSource = DefaultCameraCaptureSource(logger: ConsoleLogger(name: "CustomCameraSource"))
         self.customSource.setEventAnalyticsController(eventAnalyticsController: eventAnalyticsController)
+
+        // Create the background replacement image.
+        let rect = CGRect(x: 0,
+                          y: 0,
+                          width: self.customSource.format.width,
+                          height: self.customSource.format.height)
+        UIGraphicsBeginImageContextWithOptions(CGSize(width: self.customSource.format.width,
+                                                      height: self.customSource.format.height),
+                                               false, 0)
+        UIColor.blue.setFill()
+        UIRectFill(rect)
+        let backgroundReplacementImage: UIImage = UIGraphicsGetImageFromCurrentImageContext()!
+        UIGraphicsEndImageContext()
+
+        let backgroundReplacementConfigurations = BackgroundReplacementConfiguration(logger: ConsoleLogger(name: "BackgroundReplacementProcessor"),
+                                                                                     backgroundReplacementImage: backgroundReplacementImage)
+        self.backgroundReplacementProcessor = BackgroundReplacementVideoFrameProcessor(backgroundReplacementConfiguration: backgroundReplacementConfigurations)
+
+        let backgroundBlurConfigurations = BackgroundBlurConfiguration(logger: ConsoleLogger(name: "BackgroundBlurProcessor"),
+                                                                       blurStrength: BackgroundBlurStrength.low)
+        self.backgroundBlurProcessor = BackgroundBlurVideoFrameProcessor(backgroundBlurConfiguration: backgroundBlurConfigurations)
+
         super.init()
     }
 
@@ -69,6 +96,8 @@ class VideoModel: NSObject {
 
                 if isUsingExternalVideoSource {
                     self.customSource.removeVideoSink(sink: self.coreImageVideoProcessor)
+                    self.customSource.removeVideoSink(sink: self.backgroundBlurProcessor)
+                    self.customSource.removeVideoSink(sink: self.backgroundReplacementProcessor)
                     if isUsingMetalVideoProcessor, let metalProcessor = self.metalVideoProcessor {
                         self.customSource.removeVideoSink(sink: metalProcessor)
                     }
@@ -123,6 +152,22 @@ class VideoModel: NSObject {
         }
     }
 
+    var isUsingBackgroundBlur = false {
+        didSet {
+            if isLocalVideoActive{
+                startLocalVideo()
+            }
+        }
+    }
+
+    var isUsingBackgroundReplacement = false {
+        didSet {
+            if isLocalVideoActive {
+                startLocalVideo()
+            }
+        }
+    }
+
     private var currentRemoteVideoCount: Int {
         return remoteVideoTileStates.count
     }
@@ -153,15 +198,24 @@ class VideoModel: NSObject {
                 if self.isUsingExternalVideoSource {
                     var customVideoSource: VideoSource = self.customSource
                     customVideoSource.removeVideoSink(sink: self.coreImageVideoProcessor)
+                    customVideoSource.removeVideoSink(sink: self.backgroundBlurProcessor)
+                    customVideoSource.removeVideoSink(sink: self.backgroundReplacementProcessor)
                     if let metalVideoProcessor = self.metalVideoProcessor {
                         customVideoSource.removeVideoSink(sink: metalVideoProcessor)
                     }
+
                     if self.isUsingCoreImageVideoProcessor {
                         customVideoSource.addVideoSink(sink: self.coreImageVideoProcessor)
                         customVideoSource = self.coreImageVideoProcessor
                     } else if self.isUsingMetalVideoProcessor, let metalVideoProcessor = self.metalVideoProcessor {
                         customVideoSource.addVideoSink(sink: metalVideoProcessor)
                         customVideoSource = metalVideoProcessor
+                    } else if self.isUsingBackgroundBlur {
+                        customVideoSource.addVideoSink(sink: self.backgroundBlurProcessor)
+                        customVideoSource = self.backgroundBlurProcessor
+                    } else if self.isUsingBackgroundReplacement {
+                        customVideoSource.addVideoSink(sink: self.backgroundReplacementProcessor)
+                        customVideoSource = self.backgroundReplacementProcessor
                     }
                     self.audioVideoFacade.startLocalVideo(source: customVideoSource)
                 } else {
@@ -214,13 +268,11 @@ class VideoModel: NSObject {
                 return ComparisonResult.orderedDescending
             }
         }) as? [(Int, VideoTileState)] ?? []
-
         for remoteVideoTileState in remoteVideoStatesNotInCurrentPage {
             audioVideoFacade.pauseRemoteVideoTile(tileId: remoteVideoTileState.0)
         }
-
         if videoTilesOrderUpdated && inVideoMode {
-            videoUpdatedHandler?()
+            videoSubscriptionUpdatedHandler?()
         }
     }
 
@@ -250,16 +302,14 @@ class VideoModel: NSObject {
     }
 
     func getPreviousRemoteVideoPage() {
-        for remoteVideoTileState in remoteVideoStatesInCurrentPage {
-            audioVideoFacade.pauseRemoteVideoTile(tileId: remoteVideoTileState.0)
-        }
+        let removedList: [RemoteVideoSource] = getRemoteVideoSubscriptionsFromRemoteVideoTileStates(remoteVideoTileStates: remoteVideoStatesInCurrentPage)
+        audioVideoFacade.updateVideoSourceSubscriptions(addedOrUpdated: [:], removed: removedList)
         currentRemoteVideoPageIndex -= 1
     }
 
     func getNextRemoteVideoPage() {
-        for remoteVideoTileState in remoteVideoStatesInCurrentPage {
-            audioVideoFacade.pauseRemoteVideoTile(tileId: remoteVideoTileState.0)
-        }
+        let removedList: [RemoteVideoSource] = getRemoteVideoSubscriptionsFromRemoteVideoTileStates(remoteVideoTileStates: remoteVideoStatesInCurrentPage)
+        audioVideoFacade.updateVideoSourceSubscriptions(addedOrUpdated: [:], removed: removedList)
         currentRemoteVideoPageIndex += 1
     }
 
@@ -276,11 +326,64 @@ class VideoModel: NSObject {
             }
         }
     }
+    
+    func addAllRemoteVideosInCurrentPageExceptUserPausedVideos() {
+        var updatedSources:[RemoteVideoSource: VideoSubscriptionConfiguration] = [:]
+        let attendeeKeyMap = remoteVideoSourceConfigurations.keys.reduce(into: [String: RemoteVideoSource]()) {
+            $0[$1.attendeeId] = $1
+        }
+        let videoTilesAttendeeMap = Set(remoteVideoStatesInCurrentPage.map{ $0.1.attendeeId })
+        
+        let attendeeIds = Set(remoteVideoSourceConfigurations.keys.map { $0.attendeeId })
+        for remoteVideoTileState in remoteVideoStatesInCurrentPage{
+                let attendeeId = String(remoteVideoTileState.1.attendeeId)
+            if attendeeIds.contains(attendeeId), let key = attendeeKeyMap[attendeeId]{
+                updatedSources[key] = remoteVideoSourceConfigurations[key]
+            }
+        }
+        audioVideoFacade.updateVideoSourceSubscriptions(addedOrUpdated: updatedSources, removed: [])
+    }
+    
+    func addContentShareVideoSource(attendeeId : String) {
+        var updatedSources:[RemoteVideoSource: VideoSubscriptionConfiguration] = [:]
+        for remoteVideoSource in remoteVideoSourceConfigurations {
+            if(remoteVideoSource.key.attendeeId == attendeeId) {
+                updatedSources[remoteVideoSource.key] = remoteVideoSourceConfigurations[remoteVideoSource.key]
+                
+            }
+        }
+        audioVideoFacade.updateVideoSourceSubscriptions(addedOrUpdated: updatedSources, removed: [])
+    }
 
     func pauseAllRemoteVideos() {
         for remoteVideoTileState in remoteVideoTileStates {
             audioVideoFacade.pauseRemoteVideoTile(tileId: remoteVideoTileState.0)
         }
+    }
+    
+    func unsubscribeAllRemoteVideos() {
+        let remoteVideoSources: [RemoteVideoSource] = getRemoteVideoSubscriptionsFromRemoteVideoTileStates(remoteVideoTileStates: remoteVideoTileStates)
+        audioVideoFacade.updateVideoSourceSubscriptions(addedOrUpdated:[:], removed:remoteVideoSources)
+    }
+    
+    func getRemoteVideoSubscriptionsFromRemoteVideoTileStates(remoteVideoTileStates: [(Int, VideoTileState)]) -> [RemoteVideoSource] {
+        var remoteVideoSources: [RemoteVideoSource] = []
+        let attendeeKeyMap = remoteVideoSourceConfigurations.keys.reduce(into: [String: RemoteVideoSource]()) {
+            $0[$1.attendeeId] = $1
+        }
+        let attendeeIds = Set(remoteVideoSourceConfigurations.keys.map { $0.attendeeId })
+        for remoteVideoTileState in remoteVideoTileStates {
+                let attendeeId = String(remoteVideoTileState.1.attendeeId)
+            if attendeeIds.contains(attendeeId), let key = attendeeKeyMap[attendeeId] {
+                remoteVideoSources.append(key)
+            }
+        }
+        return remoteVideoSources
+    }
+    
+    func removeRemoteVideosNotInCurrentPage() {
+        let remoteVideoSourcesNotInCurrPage: [RemoteVideoSource] = getRemoteVideoSubscriptionsFromRemoteVideoTileStates(remoteVideoTileStates: remoteVideoStatesNotInCurrentPage)
+        audioVideoFacade.updateVideoSourceSubscriptions(addedOrUpdated:[:], removed:remoteVideoSourcesNotInCurrPage)
     }
 
     func getVideoTileState(for indexPath: IndexPath) -> VideoTileState? {
@@ -323,11 +426,62 @@ extension VideoModel: VideoTileCellDelegate {
     }
 
     func onUpdatePriorityButtonClicked(attendeeId: String, priority: VideoPriority) {
+        var updatedSources:[RemoteVideoSource: VideoSubscriptionConfiguration] = [: ]
         for (source, config) in remoteVideoSourceConfigurations {
             if attendeeId == source.attendeeId {
                 config.priority = priority
+                updatedSources[source] = config
+            }
+            
+        }
+        audioVideoFacade.updateVideoSourceSubscriptions(addedOrUpdated: updatedSources, removed: [])
+    }
+
+    func onVideoFilterButtonClicked(videoFilter: BackgroundFilter, uiView: UIViewController) {
+        switch videoFilter {
+        case .none:
+            if isUsingBackgroundBlur {
+                isUsingBackgroundBlur.toggle()
+                uiView.view.makeToast("Turning background blur off.")
+            } else if isUsingBackgroundReplacement {
+                isUsingBackgroundReplacement.toggle()
+                uiView.view.makeToast("Turning background replacement off.")
+            } else {
+                uiView.view.makeToast("No video filers are on.")
+            }
+        case .blur:
+            if isUsingMetalVideoProcessor ||
+               isUsingCoreImageVideoProcessor ||
+               isUsingBackgroundReplacement {
+                uiView.view.makeToast("Cannot toggle more than one filter at a time.")
+                return
+            }
+            let nextStatus = isUsingBackgroundBlur ? "off" : "on"
+            uiView.view.makeToast("Turning background \(videoFilter.description) \(nextStatus).")
+            isUsingBackgroundBlur.toggle()
+        case .replacement:
+            if isUsingMetalVideoProcessor ||
+               isUsingCoreImageVideoProcessor ||
+               isUsingBackgroundBlur {
+                uiView.view.makeToast("Cannot toggle more than one filter at a time.")
+                return
+            }
+            let nextStatus = isUsingBackgroundReplacement ? "off" : "on"
+            uiView.view.makeToast("Turning background \(videoFilter.description) \(nextStatus).")
+            isUsingBackgroundReplacement.toggle()
+        @unknown default:
+            self.logger.info(msg: "Unknown background filter.")
+        }
+    }
+    func onUpdateResolutionButtonClicked(attendeeId: String, resolution: VideoResolution) {
+        var updatedSources:[RemoteVideoSource: VideoSubscriptionConfiguration] = [:]
+        for(source, config) in remoteVideoSourceConfigurations {
+            if attendeeId == source.attendeeId {
+                config.targetResolution = resolution
+                updatedSources[source] = config
             }
         }
-        audioVideoFacade.updateVideoSourceSubscriptions(addedOrUpdated: remoteVideoSourceConfigurations, removed: [])
+        audioVideoFacade.updateVideoSourceSubscriptions(addedOrUpdated: updatedSources, removed: [])
     }
+    
 }